[#182] Create: Chip 폴더로 이동
